Helpers: Use nested class for better usability in Studio
